test: split two test modules at their seam
Both were over the thousand-line convention: one covers adding the shares
up and, separately, what a completion reads out of the workers own
results; the other covers the masking of a URL and, separately, what a
refusal quotes back beside it. Each half moves whole, with the module
docstrings saying where the other one is.
